Move constants to helpers folder
