+++ src/pages/components/image gallery/ImageGallery.tsx
@@ -5,7 +5,7 @@ import "react-responsive-carousel/lib/styles/carousel.min.css";
 import "./image-gallery.css";
 import "../modal.css";
 
-import { SMALL_SCREEN } from "../../../Constants";
+import { SMALL_SCREEN } from "../../../helpers/Constants";
 import { ImageGalleryCollumn } from "./ImageGalleryCollumn";
 
 function ImageGallery(props: { images: string | any[]; }) {
